chipmunk-physics: rename from chipmunk.
Needed for https://github.com/Homebrew/brew/pull/20134
